Updated busy state and home view, added build context extension
