+++ lib/ui/views/home/home_view.dart
@@ -1,10 +1,16 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+import 'package:flutter_svg/svg.dart';
 import 'package:receipe_app/generated/l10n.dart';
-import 'package:stacked/stacked.dart';
 import 'package:receipe_app/ui/common/app_colors.dart';
+import 'package:receipe_app/ui/common/app_images.dart';
 import 'package:receipe_app/ui/common/ui_helpers.dart';
+import 'package:receipe_app/ui/extension/app_typography.dart';
+import 'package:receipe_app/ui/extension/palette.dart';
+import 'package:receipe_app/ui/views/home/widgets/product_item.dart';
+import 'package:receipe_app/ui/widgets/common/app_drawer/app_drawer.dart';
+import 'package:stacked/stacked.dart';
 
-import '../../widgets/common/primary_button/primary_button.dart';
 import 'home_viewmodel.dart';
 
 class HomeView extends StackedView<HomeViewModel> {
@@ -16,69 +22,101 @@ class HomeView extends StackedView<HomeViewModel> {
     HomeViewModel viewModel,
     Widget? child,
   ) {
+    ThemeData theme = Theme.of(context);
+    AppTypography? typography = theme.extension<AppTypography>();
+    Palette? palette = theme.extension<Palette>();
+    final homeScaffoldKey = GlobalKey<ScaffoldState>();
+    
     return Scaffold(
-      body: SafeArea(
-        child: Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 25.0),
-          child: Center(
-            child: Column(
-              mainAxisSize: MainAxisSize.max,
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: [
-                verticalSpaceLarge,
-                Column(
-                  children: [
-                    const Text(
-                      'Hello, STACKED!',
-                      style: TextStyle(
-                        fontSize: 35,
-                        fontWeight: FontWeight.w900,
-                      ),
-                    ),
-                    verticalSpaceMedium,
-                    MaterialButton(
-                      color: Colors.black,
-                      onPressed: viewModel.incrementCounter,
-                      child: Text(
-                        viewModel.counterLabel,
-                        style: const TextStyle(color: Colors.white),
-                      ),
-                    ),
-                  ],
-                ),
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  children: [
-                    MaterialButton(
-                      color: kcDarkGreyColor,
-                      onPressed: viewModel.showDialog,
-                      child: const Text(
-                        'Show Dialog',
-                        style: TextStyle(
-                          color: Colors.white,
-                        ),
-                      ),
-                    ),
-                    MaterialButton(
-                      color: kcDarkGreyColor,
-                      onPressed: viewModel.showBottomSheet,
-                      child: const Text(
-                        'Show Bottom Sheet',
-                        style: TextStyle(
-                          color: Colors.white,
-                        ),
-                      ),
-                    ),
-                  ],
-                ),
-                PrimaryButton(
-                  buttonText: S.current.hello_word,
-                  onTap: () {},
-                ),
-              ],
+      key: homeScaffoldKey,
+      drawer: AppDrawer(),
+      appBar: AppBar(
+        backgroundColor: kcBackground,
+        elevation: 0,
+        leading: InkWell(
+          onTap: () {
+            homeScaffoldKey.currentState!.openDrawer();
+          },
+          child: Align(
+            alignment: Alignment.center,
+            child: Padding(
+              padding: EdgeInsets.only(
+                left: 16.w,
+                top: 8.h,
+              ),
+              child: SvgPicture.asset(
+                AppImages.burgerLog,
+                width: 24.7.w,
+                height: 24.h,
+              ),
             ),
           ),
         ),
+        title: Text(
+          S.current.home,
+          style: typography?.titleBold16?.copyWith(color: palette?.gray11),
+        ),
+      ),
+      body: SafeArea(
+        child: Column(
+          children: [
+            SizedBox(
+              height: 18.h,
+            ),
+            Builder(
+              builder: (context) {
+                // Loading
+                if (viewModel.isBusy) {
+                  return Center(
+                    child: CircularProgressIndicator.adaptive(),
+                  );
+                }
+
+                // Error
+                if (viewModel.hasError) {
+                  return Center(
+                    child: Text(viewModel.modelMessage ?? ''),
+                  );
+                }
+
+                // Empty
+                if (viewModel.data == null || viewModel.data!.isEmpty) {
+                  return Center(
+                    child: Text('There are no available dish to view currently'),
+                  );
+                }
+
+                // Loaded
+                return Expanded(
+                  child: Padding(
+                    padding: EdgeInsets.symmetric(horizontal: sidePadding),
+                    child: GridView.builder(
+                      gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
+                        crossAxisCount: 2,
+                        childAspectRatio: 0.90,
+                        mainAxisSpacing: 20.h,
+                        crossAxisSpacing: 16.w,
+                      ),
+                      itemCount: viewModel.data?.length,
+                      itemBuilder: (context, index) {
+                        final recipe = viewModel.data?.elementAt(index);
+                        return ProductItem(
+                          recipe: recipe,
+                          onTap: () => viewModel.navigateToDishDetailsView(recipe),
+                        );
+                      },
+                    ),
+                  ),
+                );
+              },
+            ),
+          ],
+        ),
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: viewModel.navigateToAddProduct,
+        backgroundColor: palette!.primary6,
+        child: SvgPicture.asset(AppImages.addcon),
       ),
     );
   }
