+++ lib/ui/widgets/common/overlay_loader/overlay_loader_model.dart
@@ -1,0 +1,3 @@
+import 'package:stacked/stacked.dart';
+
+class OverlayLoaderModel extends BaseViewModel {}
