+++ test/widget_models/overylay_loader_model_test.dart
@@ -4,16 +4,8 @@ import 'package:receipe_app/app/app.locator.dart';
 import '../helpers/test_helpers.dart';
 
 void main() {
-
-  group('HomeViewmodelTest -', () {
+  group('OverylayLoaderModel Tests -', () {
     setUp(() => registerServices());
     tearDown(() => locator.reset());
-
-    group('incrementCounter -', () {
-    });
-
-    group('showBottomSheet -', () {
-     
-    });
   });
 }
